Scraping_emploi_cm/requirements.txt
requests==2.32.4
brotli==1.1.0
beautifulsoup4==4.13.4
python-dotenv==1.1.1

